feat: add update auction

+++ server/src/facility-details/facility-details.service.ts
@@ -8,10 +8,12 @@ import {
 import { InjectRepository } from '@nestjs/typeorm';
 import { AuthErrorMessage } from 'src/auth/auth-error-message.enum';
 import { ProductTypeService } from 'src/product-type/product-type.service';
+import { ProductService } from 'src/product/product.service';
 import { UserService } from 'src/user/user.service';
 import { Repository } from 'typeorm';
 import { CreateFacilityDetailsDto } from './dto/create-facility-details.dto';
 import { UpdateFacilityDetailsDto } from './dto/update-facility-details.dto';
+import { FacilityDetailsErrorMessage } from './facility-details-error-message.enum';
 import { FacilityDetails } from './facility-details.entity';
 
 @Injectable()
@@ -22,6 +24,8 @@ export class FacilityDetailsService {
     private readonly userService: UserService,
     @Inject(forwardRef(() => ProductTypeService))
     private readonly productTypeService: ProductTypeService,
+    @Inject(forwardRef(() => ProductService))
+    private readonly productService: ProductService,
   ) {}
 
   public async create(
@@ -82,6 +86,14 @@ export class FacilityDetailsService {
 
   public async remove(id: string): Promise<void> {
     try {
+      const products = await this.productService.findAllByFacilityId(id);
+      console.log(products);
+      if (products.length !== 0) {
+        throw new HttpException(
+          FacilityDetailsErrorMessage.CannotDeleteFacilityWithProducts,
+          HttpStatus.BAD_REQUEST,
+        );
+      }
       await this.facilityDetailsRepository.delete(id);
     } catch (error) {
       throw new HttpException(error.message, HttpStatus.BAD_REQUEST);
